+++ .idea/misc.xml
@@ -14,6 +14,7 @@
         <entry key="..\:/Proyectos_Android_Kodemia/Proyecto_Final/Bankodemia-/app/src/main/res/layout/fragment_inicio.xml" value="0.2" />
         <entry key="..\:/Proyectos_Android_Kodemia/Proyecto_Final/Bankodemia-/app/src/main/res/layout/fragment_notifications.xml" value="0.25416666666666665" />
         <entry key="..\:/Proyectos_Android_Kodemia/Proyecto_Final/Bankodemia-/app/src/main/res/layout/fragment_servicios.xml" value="0.25416666666666665" />
+        <entry key="..\:/Proyectos_Android_Kodemia/Proyecto_Final/Bankodemia-/app/src/main/res/layout/fragment_tarjeta.xml" value="0.16486486486486487" />
         <entry key="..\:/Proyectos_Android_Kodemia/Proyecto_Final/Bankodemia-/app/src/main/res/menu/bottom_nav_menu.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/HP/BankodemiaApp/app/src/main/res/layout/activity_main.xml" value="0.18461538461538463" />
         <entry key="..\:/Users/HP/BankodemiaApp/app/src/main/res/layout/main_fragment.xml" value="0.16944444444444445" />
