+++ app/src/main/java/mx/kodemia/bankodemiaapp/modules/home/view/fragments/TarjetaFragment.kt
@@ -1,4 +1,4 @@
-package mx.kodemia.bankodemiaapp.ui.home.view.fragments
+package mx.kodemia.bankodemiaapp.modules.home.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import mx.kodemia.bankodemiaapp.databinding.FragmentServiciosBinding
-import mx.kodemia.bankodemiaapp.ui.home.viewmodel.ServiciosFragmentViewModel
+import mx.kodemia.bankodemiaapp.databinding.FragmentTarjetaBinding
+import mx.kodemia.bankodemiaapp.modules.home.viewmodel.TarjetaFragmentViewModel
 
-class ServiciosFragment : Fragment() {
+class TarjetaFragment : Fragment() {
 
-    private var _binding: FragmentServiciosBinding? = null
+    private var _binding: FragmentTarjetaBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -23,14 +23,14 @@ class ServiciosFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val notificationsViewModel =
-            ViewModelProvider(this).get(ServiciosFragmentViewModel::class.java)
+        val dashboardViewModel =
+            ViewModelProvider(this).get(TarjetaFragmentViewModel::class.java)
 
-        _binding = FragmentServiciosBinding.inflate(inflater, container, false)
+        _binding = FragmentTarjetaBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textNotifications
-        notificationsViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textDashboard
+        dashboardViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
