+++ app/src/main/java/mx/kodemia/bankodemiaapp/modules/home/viewmodel/InicioFragmentViewModel.kt
@@ -1,13 +1,13 @@
-package mx.kodemia.bankodemiaapp.ui.home.viewmodel
+package mx.kodemia.bankodemiaapp.modules.home.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ServiciosFragmentViewModel : ViewModel() {
+class InicioFragmentViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
