+++ app/src/main/res/layout/activity_home_details_transaction.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.view.HomeDetailsTransactionActivity">
+    tools:context=".modules.home.view.HomeDetailsTransactionActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
