+++ app/src/main/res/layout/fragment_servicios.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.view.fragments.ServiciosFragment">
+    tools:context=".modules.home.view.fragments.ServiciosFragment">
 
     <TextView
         android:id="@+id/text_notifications"
